split session in seperate file
